Container wait result type fix

+++ src/container.rs
@@ -822,7 +822,7 @@ pub struct WaitContainerResultsError {
 #[serde(rename_all = "PascalCase", deny_unknown_fields)]
 #[allow(missing_docs)]
 pub struct WaitContainerResults {
-    pub status_code: u16,
+    pub status_code: u64,
     pub error: Option<WaitContainerResultsError>,
 }
